changed addbook to now accept multi-copy creation. originals are added automatically.
modified payfine ui to accept bid so we know which borrower is paying
the fine. completed a full payfine function with multi paying support.

NOte: payments are forced in multiples of 5 and overpaying,
while it does not affect the database, is not reported. Underpaying is
supported, but there is no set order to which fines are paid off.

was modifying accountinfodialog halfway when i realized the inDate for a
borrowing is not when it is due, but when it is returned. a borrowing
entry is never deleted. updated date checks for returning books.

most borrower functions now support blocking actions if a fine is
unpaid, etc.

General Public borrower type has bookTimeLimit set to 0 temporarily for
debugging / demonstration purposes.