booktimelimit for borrower_Type changed to type int.
getbooks method now properly gets books associated with a given subject
or secondary authors. also takes in a  new parameter that tells the
method what kind of search it is. 

now connects to peter's oracle
account. search dialog modified to be compatible with getbooks.

check overdue is being worked on.